pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "politikea-ideology-mapping"
version = "0.1.0"
description = "Methodology replication toolkit for multi-axis AI-based ideological scoring of political proposals"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["political science", "ideology", "NLP", "LLM", "civic AI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas>=1.5",
    "pyarrow>=12",
    "scikit-learn>=1.2",
    "scipy>=1.10",
    "numpy>=1.24",
    "sentence-transformers>=2.2",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.1"]

[project.scripts]
ideology-mapping = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["toolkit*", "analysis*"]

[tool.ruff]
line-length = 100
target-version = "py39"
